prep/requirements.txt
openai==0.27.0
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
langchain==0.0.267
faiss-cpu==1.7.4
google-colab==0.0.1a1
anthropic==0.5.0 